+++ src/Pages/PaginaDetalhesRestaurante/PaginaDetalhesRestaurante.js
@@ -1,9 +1,9 @@
-import React, { useEffect, useContext } from 'react'
+import React, { useEffect, useContext, useState } from 'react'
 
 import GlobalContext from '../../Global/GlobalContext'
 import { useParams } from 'react-router-dom'
 import CardProduto from '../../Components/CardProduto/CardProduto'
-import { Box, Image, Flex } from '@chakra-ui/react'
+import { Box, Button, Flex, Modal, ModalBody, ModalCloseButton, ModalContent, ModalFooter, ModalHeader, ModalOverlay, Select, useDisclosure } from "@chakra-ui/react"
 import Header from '../../Components/Headers/Header'
 import BarraNavegacao from '../../Components/BarraDeNavegacao/BarraNavegacao'
 
@@ -11,7 +11,13 @@ const PaginaDetalhesRestaurante = () => {
 
   const pathParams = useParams()
 
-  const { states, requests } = useContext(GlobalContext)
+  const { states, requests, setters } = useContext(GlobalContext)
+
+  const { carrinho } = states
+  const { setCarrinho } = setters
+  const [selectedProduct, setSelectedProduct] = useState({});
+  const [valor, setValor] = useState(1)
+  const { isOpen, onOpen, onClose } = useDisclosure()
 
   const categorias = []
 
@@ -27,16 +33,69 @@ const PaginaDetalhesRestaurante = () => {
 
   // console.log(categoriasSemRepetir)
 
+
+
+  const onClickAdicionarParaCarrinho = (produto) => {
+    const novoCarrinho = [...carrinho]
+    let temNoCarrinho = false
+    for (let item of novoCarrinho) {
+      if (item.id === produto.id) {
+        item.quantity += valor
+        temNoCarrinho = true
+      }
+    }
+    if (temNoCarrinho === false) {
+      novoCarrinho.push({ ...produto, quantity: valor })
+    }
+    setCarrinho(novoCarrinho)
+    onClose()
+  }
+
+  const onClickRemoverProduto = (item) => {
+    if(item.quantity > 1){
+      const novoCarrinho = [...carrinho]
+      for(let produto of novoCarrinho){
+        if(produto.id === item.id){
+          produto.quantity -= 1
+        }
+      }
+      setCarrinho(novoCarrinho)
+    }else {
+      const novoCarrinho = carrinho.filter((produto) => {
+        return produto.id !== item.id
+      })
+      setCarrinho(novoCarrinho)
+    }
+  }
+
+  const idProdutos = []
+  const pegaId = () => {
+    states.carrinho && states.carrinho
+      .map((item) => { idProdutos.push(item.id) })
+  }
+  pegaId()
+  // console.log(idProdutos)
+
+
+  const onChangevalor = (event) => {
+    setValor(event.target.value)
+  }
+
+  const onAddProduct = (produto) => {
+    setSelectedProduct(produto)
+    onOpen()
+  }
+
   
   useEffect(() => {
     requests.pegarDetalhes(pathParams.id)
   }, [states.carrinho])
 
+
   return (
     <>
     <Header titulo="Restaurante"/>
     
-      
     <Flex
       direction='column'
       align='center'
@@ -84,11 +143,61 @@ const PaginaDetalhesRestaurante = () => {
             >
               <span>{categoria}</span>
               <Box height='1px' width='100%' bg='black' />
-              <CardProduto categoria={categoria} />
+
+              {states.detalhes.products && states.detalhes.products
+                .filter((produto) => {
+                  return produto.category === categoria
+                })
+                .map((produto) => {
+                  return (
+                    <CardProduto
+                      key={produto.id}
+                      id={produto.id}
+                      photoUrl={produto.photoUrl}
+                      name={produto.name}
+                      description={produto.description}
+                      price={produto.price}
+                      produtos={states.detalhes.products}
+                      valor={
+                        idProdutos.includes(produto.id) ?
+                        (states.carrinho.quantity) : ""
+                      }
+                      idProdutos={idProdutos}
+                      onClickRemoverProduto={() => onClickRemoverProduto(produto)}
+                      onAddProduct={() => onAddProduct(produto)}
+                    />
+                  )})
+              }
             </Flex>
           )
         })}
       </Flex>
+      <Modal isOpen={isOpen} onClose={onClose} isCentered >
+        <ModalOverlay />
+        <ModalContent>
+          <ModalHeader>Selecione a quantidade desejada</ModalHeader>
+          <ModalCloseButton />
+          <ModalBody>
+            <Select value={valor} onChange={onChangevalor}>
+              <option value={1}>1</option>
+              <option value={2}>2</option>
+              <option value={3}>3</option>
+              <option value={4}>4</option>
+              <option value={5}>5</option>
+              <option value={6}>6</option>
+              <option value={7}>7</option>
+              <option value={8}>8</option>
+              <option value={9}>9</option>
+              <option value={10}>10</option>
+            </Select>
+          </ModalBody>
+          <ModalFooter>
+            <Button color='#5CB646' variant='ghost' mr={3} onClick={() => onClickAdicionarParaCarrinho(selectedProduct)}>
+              ADICIONAR AO CARRINHO
+            </Button>
+          </ModalFooter>
+        </ModalContent>
+      </Modal>
 
     </Flex>
     <BarraNavegacao />
